Split model extractor test (i.e. high memory) into seperate run
