+++ .github/workflows/python-package.yml
@@ -42,6 +42,9 @@ jobs:
         python -m spacy download en_core_web_sm
         python -m pliers.support.setup_yamnet
     - name: Test with pytest
+      env:
+        skip_high_memory: true
       run: |
         py.test pliers/tests --cov=pliers --cov-report xml -m "not requires_payment" -W ignore::UserWarning
+        skip_high_memory=false py.test pliers/tests/extractors/test_model_extractors.py --forked --cov-append --cov=pliers --cov-report xml -m "not requires_payment" -W ignore::UserWarning
     - uses: codecov/codecov-action@v1
