+++ pliers/tests/extractors/test_misc_extractors.py
@@ -72,7 +72,8 @@ def test_metric_extractor():
     assert r_lambda_df['custom_function'][0] == -4
 
 
-@pytest.mark.skipif(environ.get('CI', False) == 'true', reason='high memory')
+# Change test to use simpler extractor
+@pytest.mark.skipif(environ.get('skip_high_memory', False) == 'true', reason='high memory')
 def test_metric_er_as_stim():
     stim = ComplexTextStim(text = 'This is [MASK] test')
     ext_bert = BertLMExtractor(return_softmax=True)
